Updated money to use BigDecimal instead of float

+++ services/paymentGateway/src/main/java/org/bazaar/paymentGateway/payment/dto/UserAccountRequest.java
@@ -1,5 +1,7 @@
 package org.bazaar.paymentGateway.payment.dto;
 
+import java.math.BigDecimal;
+
 import org.bazaar.paymentGateway.constant.ValidationMessage;
 
 import jakarta.validation.constraints.Email;
@@ -7,8 +9,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Positive;
 
 public record UserAccountRequest(
-        @Email(message = ValidationMessage.INVALID_EMAIL, regexp = ValidationMessage.EMAIL_REGEX) String email,
-        @NotBlank(message = "password" + ValidationMessage.NOT_BLANK) String password,
-        @Positive(message = "amountOfMoney" + ValidationMessage.POSITIVE) float amountOfMoney) {
+                @Email(message = ValidationMessage.INVALID_EMAIL, regexp = ValidationMessage.EMAIL_REGEX) String email,
+                @NotBlank(message = "password" + ValidationMessage.NOT_BLANK) String password,
+                @Positive(message = "amountOfMoney" + ValidationMessage.POSITIVE) BigDecimal amountOfMoney) {
 
 }
